Make PkgLength return something useful
Because we constantly shave bits off the stream as we successfully parse
stuff, we can't use the raw PkgLength for anything useful. Instead, we
work out when in the stream we should stop parsing, and provide that through
utility methods on `PkgLength`. We still need to be able to parse raw
lengths for the tests though, so we don't have to constuct streams that are
thousands of bytes long for the longer lengths.