Allow access to the AmlContext throughout the parse
This should make it super easy to e.g. add stuff to the namespace, or
change the current scope. In the future, it will mean we can set and access
args and locals when we actually execute AML methods.

It does make the combinators more ugly, which bleeds into the imperatively
written parsers (e.g. raw_pkg_length). In the future, I hope to rewrite
these parsers to make use of more powerful combinators, but I haven't worked
out how yet.