Parse the start of MethodInvocation (hopefully)
This is a super unclear part of the spec, so I'm not sure whether this is
even correct (but it does appear to work as intended with the QEMU tables).
It seems incorrect to treat non-method objects as invocable by a
MethodInvocation, but I can't think what else must parse these objects.