Finally properly handle error propagation in or and choice!
This was surprisingly easy to get working in the new parser, and fixes the
bad diagnostics we were previously seeing, as well as enabling us to use
the error system to propagate return values of control methods. It also
allows us to fix some bad unit tests that had to accept the "wrong" error
being produced due to use of `choice!`, which should appear as an
implementation detail but wasn't.