Move to new namespace system
We were outgrowing the old namespace system, and need a way to refer to
values without actually borrowing them (this is provided by the new concept
of an `AmlHandle`). For the meantime, this representation is slightly less
efficient (because we're not using handles yet, and need two maps), but in
the future will be much better because we'll be able to throw handles around
to avoid having to clone objects. We can also build a more efficient
representation than the BTreeMap to store the actual name part of the
namespace (the `name_map` field atm) so that it uses far less memory, and
may even be faster to search.