Changing ACPI spec link to the web page with the latest version
I noticed that the link to the ACPI specification was pointing to a specific version (6.2A) that is out of date. Changing the link to point to the UEFI web page where the latest version of the spec is posted. As of the date of this commit, the current version is 6.3. 

Kevin Shaw, Technical Writer, Intel Corporation