Start work on testing framework
This adds the start of a program to automatically compile and test a
bunch of ASL files against out parser. It relies on `iasl` being
installed to compile the files. It's still very basic and not polished
at all, but will work as a MVP until it's finished and I can be bothered
to polish it a bit.