Return both resolved path and handle when we search the namespace
Sometimes, we need to construct another path to a child object after
finding something in the namespace, and so we need the full resolved name
instead of the handle to the object itself. We have this easily available
during the search, so we might as well provide it to the caller.