Remove dependency on typenum and make safer
Replaces the dependency on typenum with const generics. Always nice!

Secondly, changes the representation of "potentially-there" fields to be
safer. The old version clearly didn't miscompile, but I don't think it was
very safe. This explicitly treats the extended fields as uninitialized
memory until a sufficient version is provided, which I think is sound?