Store a handler as a trait object in the AmlContext
This doesn't really need to be heap allocated, but I think this is a good
compromise because keeping a reference to the handler would add both another
lifetime and a type parameter to AmlHandler and therefore every single type
that refers to AmlHandler. I think we can swallow the cost of a dynamic
dispatch here.