Take the context on value-conversion operations
This is slightly horrifying, but it seems that some implicit conversions
can involve things like reading from PCI config space (e.g. converting a
FieldUnit into an Integer in certain cases). I'm not sure this is 100%
correct just from how ridiculous it sounds, but it does appear to produce
the correct results...