Move to allowing collection of tables without dispatch
This is at the heart of the new architecture - we now collect all the ACPI
tables without dispatching them ourselves. This makes the library more
flexible, at the cost of an extra step if you don't care about which tables
you have, but it's turned out much cleaner than I'd hoped overall.

Unfortunately the diff algorithm has butchered this, so this is kinda messy.