Move interrupt and platform info over to new system
One of the problems with the new architecture is that to get a nice view
of the topology and interrupt model, we need to get a bunch of "unrelated"
info out of a couple of the tables. We encapsulate this in `PlatformInfo`,
which amalgamates the FADT and MADT into some common info.