Fix first fuzzing bug! Don't allow NullName as a full NameString
The spec isn't clear about this, but in the library we assume that an
`AmlName` is not empty. However, if a NullName appears as the only
element of a NamePath, we can accidently create an empty AmlName when we
parse a NameString. This has never come up in real tables, but was detected
during fuzzing as breaking a whole bunch of stuff in the namespace searching
(due to the assumptions about AmlName being broken).