Don't panic when working with invalid PkgLengths (round 2)
The fuzzer still manages to crash the parser with an invalid PkgLength. We
were already checking that it's valid when we construct it, so we're adding
another check when we try and use it. We should look at this more closely
in the future, as I don't think we should necessarily have to do both.