Handle buffers that don't initialize all elements correctly
We were originally confused about what we should do if a buffer's data
doesn't contain the same number of elements as the supplied size, and the
spec really doesn't make it easy to work out. I think we should create the
buffer at the specified size and initialize all remaining elements to zero.
This makes buffers nicer to work with in a few places.