Improve code generation of choice! macro
This improves the code generation for when we need to try a bunch of parsers
and see what sticks. I don't know how it took me so long to think of doing
this, but it helps a huge amount. Compile times are noticeably reduced on
my machine, and we no longer need to manage the type limit when choice!ing
a large number of parsers together. The code generation at the macro
callsite is arguably worse, but overall it's a big win.