Remove centralized ACPI versions
Looking at this closer, it doesn't look like this is as simple as it looked
before. Each table's revision fields seem to be incremented as needed, as
opposed to being linked to an actual ACPI version. We hard-code the FADT
revision to do the same thing there (NOTE: I'm not sure that the FADT
should actually be using this mechanism - "if the field is present" points
me towards just checking the length, instead of the revision?).

NOTE: this module isn't public, so this doesn't break semver