Fix bug in extended field handling in the FADT
Before, we would report the existance of these optional fields if the
extended version of the field exists, even if its zero. For all of the
fields, if the extended field is zero (it's not entirely clear whether
we need to check the entire structure, or just the address?), we drop back
to the original field.