Remove AML legacy mode, and don't pre-register scopes
I don't know why we thought we needed to pre-register scopes, they will be
created normally as devices are created.

I also don't think the "legacy mode" thing is a good way of handling old
tables. I think we should instead just be aware of the weird intracacies
(e.g. DefProcessor objects), as I don't think it'll be easy to work out
if you want to set legacy mode or not.