Introduce `Propagate` type to parser
We need slightly more advanced logic in AML than all propagations being
errors. Until now, we've modelled these reasons (e.g. returning a value
from a method) as "fake" errors, but this is hacky and incorrect for a
reason we'll have to correct soon. Let's fix it for good.

This commit will not build, and should be taken as a unit with the next
commits.