Use new Propagate::Return variant to return values
This removes the fake error used to return values by abusing the propagation
of errors. This is required for the whole reason for finally getting rid of
this hack - allowing `AmlError` to remain `Eq` (which is very useful), while
allowing `AmlValue` to not implement those traits (which it shouldn't; that
change will be made in a future commit).