Remove PartialEq and Eq from AmlValue
It is not correct for AmlValue to be PartialEq or Eq, as to correctly compare
two AML objects, various conversion rules need to be applied, if they are
appropriate. This requires the context, and so is implemented with a custom
`cmp` method.

This is why we can't hold `AmlValue`s in `AmlError`, requiring the change
to use `Propagate`.

Unfortunately, this basically breaks our testing framework. I've made the
changes needed to fix all of this, but it's a bit nasty, and would probably
benefit from some cleaning up in the future.