Record attempted conversion in error
When a conversion error occurs, it is quite useful to know exactly which
conversion failed. By storing both the current value and the target
value in the `AmlError::IncompatibleValueConversion` enum variant, this
information can be provided to the caller. Additionally, by using a
struct enum variant, there is no need to keep track of which value is
the current value, and which is the target value, as this is encoded in
the type.