Regression: `AmlContext: Send + Sync` again.
Before the introduction of native methods, the `AmlContext` structure was
both `Send` and `Sync`, allowing it to be stored inside containers that
require this such as `spin::Once`. I use this to allow the global structure to
be shared without reparsing the AML.

To do this it makes two changes:

1. The `Fn` for native methods in `MethodCode::Native` is now required to
be `Send` and `Sync`.
2. It now wraps the native function in an `Arc` instead of `Rc`