Reference count inner data of AML buffer values
The inner data of buffers can be held and modified by both Buffer objects,
but also BufferField objects. I wonder if a good way of solving this neatly
would actually be to reference count this inner data, and hold Arcs to it
in both the parent buffer and any fields pointing into that buffer - let's
find out.