Parse DefCreateBitField
This is our first working buffer field implementation, and can serve as
the foundation for the rest of them pretty easily, hopefully. It seems to
work correctly, but we'll need to implement support for reading from and
writing to buffer fields manually. Lukily, it errors at the moment, so it
won't confuse anyone to add this.