Implement DefWhile
This was slightly more involved than lots of the other opcodes (as expected),
as we need to parse part of the stream potentially many times. I thought
we'd have to introduce new versions of a bunch of parsers that don't fully
evaluate the result of a parse, or introduce a new evaluation phase into
existing parsers, but this finds a way around that by allowing the 'extraction'
of the raw AML stream that another parser streams, allowing it to be reparsed.
This doesn't feel amazingly elegant, but it's fairly robust and works well
enough.