Parse MADT entries related to the X2APIC
We don't actually need to, and probably can't reliably, separate the APIC
and X2APIC at the ACPI level, as legacy OSs require X2APICs with UIDs less
than 256 to be defined with normal LAPIC entries, and so some OEMs decide
to ship tables that do this. The OSPM will need to detect support for the
X2APIC separately and then deal with this info accordingly.