Wrap buffer data in spinning_top::Spinlock
This is a stepping stone / first attempt at making it possible to mutate
underlying buffer data, which requires getting access to data within the
Arc. While totally not ideal to be using a spinlock here, it works for
now and allows us to get writes to buffer fields working.

We've got a few ideas on how to improve this in the future, but for now I
think this is fine.