aml: fix breakage by not deriving `Debug` on `AmlContext`
`AmlContext` can't derive `Debug` because we don't require the `Handler`
to be `Debug`. We could do this with a custom implementation but I don't
think there's much point; the state of the context is not super useful when
you can just access the children.

Unfortunately this change was made by a PR that clearly was not tested to
compile (irritating), and then not caught by CI because it broke months ago
when we renamed `master` to `main` :(