aml: update to `bitvec v1.0.1` as `funty v1.2` has been yanked
Fixes #133

This currently just patches our use of `bitvec` in `aml`, but I wonder if
this could be done more ergomically by actually exposing our use of `BitVec`
to represent buffers and their fields in `aml` (i.e. use it as the actual
representation of `AmlValue::Buffer`?) Maybe need to read the spec for the
exact behaviour of fields etc. to see if this would work well.