Add hack to get `aml` compiling on 32-bit platforms
This is a, hopefully very temporary, hack to get `aml` to compile on 32-bit
platforms. The issue is with `bitvec`, which seems to be unable to operate
on 64-bit values on 32-bit systems, which I don't quite understand atm. In
the long term, we should sit down and understand how to use `bitvec`
properly, or replace it with a simple implementation without this limitation.