acpi: Allow Madt and Mcfg fields to be accessed without allocator_api
In the absence of the allocator_api feature being enabled and/or other situtations it is useful to be able to access ACPI table fields directly.

This also allows the low-level use of the tables, especially when higher-level processing is not yet implemented.

Thus, this revision makes them and the Mcfg 'entries()' method public.