Re-merge `RegionMapper` and `Handler` into a single trait
This was unneeded complexity in the end and came unstuck when we realised
the library needs to do I/O port accesses for register accesses. Users can
just not implement parts of the API they're unable/will not use and deal with
the consequences.

Also catches changes to cache a mapping to the FACS in the AML interpreter.